Split video, localization, and SEO cockpit helpers
