+++ frontend/app/(core)/admin/seo/cockpit/_components/SeoCockpitRangeTabs.tsx
@@ -1,0 +1,29 @@
+import Link from 'next/link';
+import type { GscRangeKey } from '@/lib/seo/gsc-analysis';
+
+const RANGE_OPTIONS: Array<{ value: GscRangeKey; label: string }> = [
+  { value: '7d', label: '7 days' },
+  { value: '28d', label: '28 days' },
+  { value: '3m', label: '3 months' },
+];
+
+export function RangeTabs({ current, basePath }: { current: GscRangeKey; basePath: string }) {
+  return (
+    <div className="inline-flex rounded-lg border border-hairline bg-bg p-1">
+      {RANGE_OPTIONS.map((option) => (
+        <Link
+          key={option.value}
+          href={`${basePath}?range=${option.value}`}
+          prefetch={false}
+          className={
+            option.value === current
+              ? 'rounded-md bg-slate-950 px-3 py-1.5 text-xs font-semibold text-white'
+              : 'rounded-md px-3 py-1.5 text-xs font-semibold text-text-secondary hover:bg-surface-hover hover:text-text-primary'
+          }
+        >
+          {option.label}
+        </Link>
+      ))}
+    </div>
+  );
+}
